requirements.txt
anthropic>=0.40.0
replicate>=0.34.0
requests>=2.31.0
Pillow>=10.0.0
